pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "engram"
version = "0.1.0"
description = "Persistent memory for AI coding agents"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "watchdog>=4.0",
    "jinja2>=3.1",
    "filelock>=3.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
engram = "engram.cli:cli"

[tool.setuptools.packages.find]
include = ["engram*"]
